added placeholder image, and styling edit profile
